[ovsp4rt_v3] Extract DoConfigSrcPortEntry
