+++ ovs-p4rt/sidecar/core/v3/CMakeLists.txt
@@ -13,6 +13,7 @@ target_sources(ovs_sidecar_o PRIVATE
 if(ES2K_TARGET)
   target_sources(ovs_sidecar_o PRIVATE
       do_config_ip_mac_map_entry.cc
+      do_config_src_port_entry.cc
       do_config_vlan_entry.cc
   )
 endif()
