+++ ovs-p4rt/sidecar/core/v3/do_config_ip_mac_map_entry.cc
@@ -8,8 +8,6 @@
 #error "ASSERT: Target type must be ES2K_TARGET"
 #endif
 
-#include <absl/status/status.h>
-#include <absl/status/statusor.h>
 #include <arpa/inet.h>
 
 #include "client/ovsp4rt_client_interface.h"
