+++ ovs-p4rt/sidecar/core/v3/do_config_tunnel_entry.cc
@@ -4,6 +4,8 @@
 
 // Implements DoConfigTunnelEntry() (dpdk, es2k)
 
+#include <arpa/inet.h>
+
 #include "client/ovsp4rt_client_interface.h"
 #include "core/api/ovsp4rt_internal_api.h"
 #include "core/common/ovsp4rt_encoders.h"
